+++ changelog.md
@@ -2,6 +2,7 @@
 
 * Bump to Apache TinkerPop 3.3.3
 * Improved syntax for specifying `VertexProperty$Cardinality` by adding `(cardinality :single)`.
+* Improved syntax for specifying `Column` by adding `(column :values)`.
 
 ## Changes in Ogre 3.3.2.0